ptp_ocp: Add I2C bus support
Add support for the I2C IMU and OSC busses, which are provided by
xilinx xiic-2c driver.  This driver wants to claim the PCI resources
by itself, so the OCP driver cannot claim the entire BAR.

Restructure the driver so each resource is claimed and mapped
separately, delegating sub-resources to individual drivers.

The resources can be seen in the sysfs tree, e.g.:

   /sys/bus/pci/drivers/ptp_ocp/0000:12:00.0/
       ptp/ptp9
       tty/ttyS2
       tty/ttyS3
       xiic-i2c.0
       xiic-i2c.1