ptp_ocp: add SPI devices
The new bitstream adds SPI controllers, so hook them up.  There is
also an HWICAP controller, which hasn't been handled yet.

The irq vector size has increased to 10 in order to handle the
new devices.  Rework the device registration a bit so devices are
registered or skipped based on the irq availability.