ptp_ocp: restructure driver for better resource assignments.
Rework the ocp resources, removing the fixed resource entries in
favor of a dynamic table.  The table is now attached to the PCI
device id, paving the way for different resource entries.

Hook up the SPI flash.  While still somewhat flaky, reads from the
flash are successful.  Writes are not working yet.