Update for firmware v13.
Another timestamper is added, which timestamps the PHC.  This
may be used for a few things:

 - if the PHC is hardware looped back to another timestamper,
   this can accurately measure the cable delay.
 - the timestamp interrupt may be sent to the kernel, creating
   a pps device.

From within the /sys/class/timecard/ocpN directory:
 - "echo 1 > ptp/pps_enable" turns on the PPS device
 - "cat pps/assert" shows the current PPS stamp

adjtime now attempts to fold in the ns adjustment rather than
just overwriting the previous value.

The PTP OFFSET_EXTENDED ioctl works by taking a system timestamp
before and after obtaining the PHC reading.  The midpoint of the
window is considered to be the matching system time for the PHC
timestamp.

If the window is inaccurate (too large), then the midpoint never
matches up with the correct PHC time, creating a phase error.

Try and tighten up the window bounds (reported in the debug summary)
by skipping some unnecessary PCIe reads.

Add a pci_delay attribute which pulls in the trailing window edge,
indirectly adjusting the midpoint.  This isn't ideal, and is
subject to change.

Resource failures are now ignored.

NMEA default baud rate is changed to 115200