Initial checkin for eval board bench setup of PTP over UWB
This is the initial check in for UWB based synchronization efforts.

The setup has 4 main components
1. A Raspberry Pi Pico as the controller.
2. A AD9546 DPLL Eval board
3. A Decawave DW1000 eval board
4. A AD9513 Eval board, to convert the DPLL output to CMOS for the Decawave

The Pico configures and controls the DPLL over the serial pins on the eval board. It loads config files generated by the eval board GUI, and converted into C byte-arrays by the python script.

The Pico controls a Decawave DW1000 UWB IC over SPI. One device is designated as the clock source (GUG) , and the other is the clock follower (Time Stick). There are two pins on the Pico that detect whether to act as a GUG or a Time Stick.

Using UWB, a protocol similar to PTP using Sync+Followup Messages broadcasts time from the GUG to any Time Stick. The Time Stick then uses these messages to adjust the DPLL for synchronization and syntonization.

This current code can achieve synchronization, but syntonization is still being developed. This current setup is precarious due to the number of wires and current draw being sourced from the Pico, and future check-ins will be for a better hardware setup.